+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+* CXV1-6119 - Fixed media list add btn disabling issue
+* CXV1-6120 - Fixed issue with Submit button not enabling on media lists
 * CXV1-6027 - Enable drag-sorting for media list items
 * CXV1-4953 - Created App Dock for interaction data from interactions table
 * CXV1-6012 - Added dashboard categories so they can be grouped together in dropdown